refactor: 기존에 일일이 관리하던 state를 useForm으로 관리하도록 수정
