+++ src/apis/points/request.ts
@@ -1,0 +1,13 @@
+export interface SearchPointOptionsRequest {
+    point_option_name: string;
+  }
+
+export interface PointOptionRequest {
+    score: number,
+    name: string,
+  }
+
+export interface PointOptionUnderBarRequest {
+    score_: number,
+    name_: string,
+  }
