+++ src/components/modals/GivePointOptionsModal.tsx
@@ -10,7 +10,7 @@ import {
 import { useState } from 'react';
 import styled from 'styled-components';
 import { useAddPointOption, useGivePointOption } from '@/apis/points';
-import { SearchPointOptionsRequest } from '@/apis/points/response';
+import { PointOptionRequest, SearchPointOptionsRequest } from '@/apis/points/request';
 import { useDropDown } from '@/hooks/useDropDown';
 import { useForm } from '@/hooks/useForm';
 import { usePointOptionList } from '@/hooks/usePointsApi';
@@ -21,26 +21,27 @@ interface PropsType {
   close: () => void;
 }
 
-const canClick = true;
 export function GivePointOptionsModal({ close, selectedStudentId }: PropsType) {
+  const canClick = true;
+
   const [newItem, setNewItem] = useState(true);
+  const [selectedPointOption, setSelectedPointOption] = useState<string>('');
+
   const { onDropDownChange, sort } = useDropDown<string>('');
-  const [addPointOption, setAddPointOption] = useState({
-    score: 0,
-    name: '',
-  });
+
+  const { state: pointOptionState, onHandleChange: setPointOptionState } =
+    useForm<SearchPointOptionsRequest>({
+      point_option_name: '',
+    });
+
+  const { state: addPointOption, onHandleChange: setAddPointOption } =
+    useForm<PointOptionRequest>({
+      score: 0,
+      name: '',
+    });
 
   const { score: scoreOption, name: nameOption } = addPointOption;
 
-  const onAddPointOption = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { value, name: option } = e.target;
-    setAddPointOption({
-      ...addPointOption,
-      [option]: value,
-    });
-  };
-
-  const [selectedPointOption, setSelectedPointOption] = useState<string>('');
   const { data: allPointOptions } = usePointOptionList();
 
   const newItemInput = () => {
@@ -50,11 +51,6 @@ export function GivePointOptionsModal({ close, selectedStudentId }: PropsType) {
   const onClickPointOption = (id: string) => {
     setSelectedPointOption((OptionId) => (OptionId === id ? '' : id));
   };
-
-  const { state: pointOptionState, onHandleChange } =
-    useForm<SearchPointOptionsRequest>({
-      point_option_name: '',
-    });
 
   const givePointOptionAPI = useGivePointOption(
     selectedPointOption,
@@ -93,7 +89,7 @@ export function GivePointOptionsModal({ close, selectedStudentId }: PropsType) {
           placeholder="ex) 봉사활동"
           name="point_option_name"
           value={pointOptionState.point_option_name}
-          onChange={onHandleChange}
+          onChange={setPointOptionState}
         />
       </_SearchWrapper>
       <_PointOptionList className="grantPoint">
@@ -105,6 +101,7 @@ export function GivePointOptionsModal({ close, selectedStudentId }: PropsType) {
             const { point_option_id, name, type, score } = options;
             return (
               <PointItem
+                key={point_option_id}
                 point_history_id={point_option_id}
                 name={name}
                 type={type}
@@ -133,9 +130,7 @@ export function GivePointOptionsModal({ close, selectedStudentId }: PropsType) {
         )}
       </_AddImgWrapper>
       <_AddInputBigWrapper className="grantPoint">
-        {newItem ? (
-          ''
-        ) : (
+        {!newItem && (
           <Input
             className="grantPoint"
             width={478}
@@ -144,13 +139,11 @@ export function GivePointOptionsModal({ close, selectedStudentId }: PropsType) {
             placeholder="ex) 무단 외출"
             name="name"
             value={nameOption}
-            onChange={onAddPointOption}
+            onChange={setAddPointOption}
           />
         )}
         <_AddInputSmallWrapper className="grantPoint">
-          {newItem ? (
-            ''
-          ) : (
+          {!newItem && (
             <Input
               className="grantPoint"
               width={243}
@@ -159,12 +152,10 @@ export function GivePointOptionsModal({ close, selectedStudentId }: PropsType) {
               placeholder="ex) 12 (숫자만 입력)"
               name="score"
               value={scoreOption}
-              onChange={onAddPointOption}
+              onChange={setAddPointOption}
             />
           )}
-          {newItem ? (
-            ''
-          ) : (
+          {!newItem && (
             <DropDown
               className="grantPoint"
               width={216}
