+++ src/components/modals/ViewPointOptionsModal.tsx
@@ -11,16 +11,13 @@ import { Dispatch, SetStateAction, useState } from 'react';
 import styled from 'styled-components';
 import {
   useAddPointOption,
-  useDeletePointOption,
   useEditPointOption,
 } from '@/apis/points';
-import { SearchPointOptionsRequest } from '@/apis/points/response';
+import { PointOptionRequest, PointOptionUnderBarRequest, SearchPointOptionsRequest } from '@/apis/points/request';
 import { useDropDown } from '@/hooks/useDropDown';
 import { useForm } from '@/hooks/useForm';
 import { usePointOptionList } from '@/hooks/usePointsApi';
 import { PointItem } from '../main/DetailBox/PointItem';
-import { useModal } from '@/hooks/useModal';
-import { DeletePointOptionModal } from './DeletePointOption';
 
 interface PropsType {
   selectedPointOption?: string;
@@ -28,45 +25,36 @@ interface PropsType {
   close: () => void;
 }
 
-const MustTrue = true;
 
 export function ViewPointOptionsModal({
   close,
   selectedPointOption,
   setSelectedPointOption,
 }: PropsType) {
+  
+  const MustTrue = true;
+
   const [newItem, setNewItem] = useState(true);
+
   const { onDropDownChange: AddChange, sort: AddState } =
     useDropDown<string>('');
   const { onDropDownChange: EditChange, sort: EditState } =
     useDropDown<string>('');
-  const [addPointOption, setAddPointOption] = useState({
-    score: 0,
-    name: '',
-  });
-  const [editPointOption, setEditPointOption] = useState({
-    score_: 0,
-    name_: '',
-  });
+
+  const { state: addPointOption, onHandleChange: setAddPointOption } =
+    useForm<PointOptionRequest>({
+      score: 0,
+      name: '',
+    });
+  const { state: editPointOption, setState: setEditPointOption, onHandleChange: EditPointOptionHandler } =
+    useForm<PointOptionUnderBarRequest>({
+      score_: 0,
+      name_: '',
+    });
 
   const { score: addPointScore, name: addPointName } = addPointOption;
   const { score_, name_ } = editPointOption;
 
-  const onAddPointOption = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { value, name } = e.target;
-    setAddPointOption({
-      ...addPointOption,
-      [name]: value,
-    });
-  };
-
-  const onEditPointOption = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { value, name } = e.target;
-    setEditPointOption({
-      ...editPointOption,
-      [name]: value,
-    });
-  };
 
   const { data: allPointOptions } = usePointOptionList();
 
@@ -180,7 +168,6 @@ export function ViewPointOptionsModal({
       {selectedPointOption ? (
         <>
           <_AddImgWrapper
-            onClick={selectedPointOption ? undefined : newItemInput}
             newItem={newItem}
           >
             <_Text>항목 수정</_Text>
@@ -194,7 +181,7 @@ export function ViewPointOptionsModal({
               placeholder="ex) 무단 외출"
               name="name_"
               value={name_}
-              onChange={onEditPointOption}
+              onChange={EditPointOptionHandler}
             />
             <_AddInputSmallWrapper>
               <Input
@@ -204,7 +191,7 @@ export function ViewPointOptionsModal({
                 placeholder="ex) 12 (숫자만 입력)"
                 name="score_"
                 value={score_}
-                onChange={onEditPointOption}
+                onChange={EditPointOptionHandler}
               />
               <DropDown
                 width={216}
@@ -233,9 +220,7 @@ export function ViewPointOptionsModal({
             )}
           </_AddImgWrapper>
           <_AddInputBigWrapper>
-            {newItem ? (
-              ''
-            ) : (
+            {!newItem && (
               <Input
                 className="grantPoint"
                 width={478}
@@ -244,13 +229,11 @@ export function ViewPointOptionsModal({
                 placeholder="ex) 무단 외출"
                 name="name"
                 value={addPointName}
-                onChange={onAddPointOption}
+                onChange={setAddPointOption}
               />
             )}
             <_AddInputSmallWrapper>
-              {newItem ? (
-                ''
-              ) : (
+              {!newItem &&  (
                 <Input
                   className="grantPoint"
                   width={243}
@@ -259,12 +242,10 @@ export function ViewPointOptionsModal({
                   placeholder="ex) 12 (숫자만 입력)"
                   name="score"
                   value={addPointScore}
-                  onChange={onAddPointOption}
+                  onChange={setAddPointOption}
                 />
               )}
-              {newItem ? (
-                ''
-              ) : (
+              {!newItem && (
                 <DropDown
                   className="grantPoint"
                   width={216}
